I am working on fixing the eevdf compiler errors.

+++ src/system/kernel/scheduler/HashMap.h
@@ -50,7 +50,7 @@ public:
 
             for (uint32 i = 0; i < fBucketCount; i++) {
                 if (fBuckets[i].key != 0) {
-                    uint32 newHash = (uintptr_t)fBuckets[i].key % newBucketCount;
+                    uint32 newHash = (uint32)fBuckets[i].key % newBucketCount;
                     while (newBuckets[newHash].key != 0) {
                         newHash = (newHash + 1) % newBucketCount;
                     }
@@ -63,7 +63,7 @@ public:
             fBucketCount = newBucketCount;
         }
 
-        uint32 hash = (uintptr_t)key % fBucketCount;
+        uint32 hash = (uint32)key % fBucketCount;
         while (fBuckets[hash].key != 0) {
             hash = (hash + 1) % fBucketCount;
         }
